Reorganize typeDefs and resolvers into separate files
