+++ src/schema/team.js
@@ -1,0 +1,16 @@
+import { makeExecutableSchema, addMockFunctionsToSchema } from 'graphql-tools';
+
+let typeDefs = `
+type Team {
+  id: ID!
+  dreamId: ID!
+  title: String
+  url: String!
+}
+`;
+
+let schema = makeExecutableSchema({ typeDefs, resolvers });
+
+addMockFunctionsToSchema({ schema, preserveResolvers: true });
+
+export default schema;
